requirements.txt
flask==3.0.3
werkzeug==3.0.3
python-dotenv==1.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
blinker==1.7.0
google-generativeai==0.3.2
flask-cors==4.0.0